Package rename / pub fixed size
