+++ core/src/worker/slot_supplier.rs
@@ -1,56 +1,1 @@
-use std::{marker::PhantomData, sync::Arc};
-use temporal_sdk_core_api::worker::{
-    SlotKind, SlotReservationContext, SlotSupplier, SlotSupplierPermit,
-};
-use tokio::sync::Semaphore;
 
-mod resource_based;
-
-pub use resource_based::{
-    RealSysInfo, ResourceBasedSlots, ResourceBasedTuner, ResourceSlotOptions,
-};
-
-pub(crate) struct FixedSizeSlotSupplier<SK> {
-    sem: Arc<Semaphore>,
-    _pd: PhantomData<SK>,
-}
-
-impl<SK> FixedSizeSlotSupplier<SK> {
-    pub(crate) fn new(size: usize) -> Self {
-        Self {
-            sem: Arc::new(Semaphore::new(size)),
-            _pd: Default::default(),
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl<SK> SlotSupplier for FixedSizeSlotSupplier<SK>
-where
-    SK: SlotKind + Send + Sync,
-{
-    type SlotKind = SK;
-
-    async fn reserve_slot(&self, _: &dyn SlotReservationContext) -> SlotSupplierPermit {
-        let perm = self
-            .sem
-            .clone()
-            .acquire_owned()
-            .await
-            .expect("inner semaphore is never closed");
-        SlotSupplierPermit::with_user_data(perm)
-    }
-
-    fn try_reserve_slot(&self, _: &dyn SlotReservationContext) -> Option<SlotSupplierPermit> {
-        let perm = self.sem.clone().try_acquire_owned();
-        perm.ok().map(SlotSupplierPermit::with_user_data)
-    }
-
-    fn mark_slot_used(&self, _info: SK::Info<'_>) {}
-
-    fn release_slot(&self) {}
-
-    fn available_slots(&self) -> Option<usize> {
-        Some(self.sem.available_permits())
-    }
-}
